Lift out common components
