+++ client/packages/system/src/Clinician/ClinicianSearchInput.tsx
@@ -5,14 +5,21 @@ import {
   IconButton,
   PlusCircleIcon,
   Typography,
+  useConfirmationModal,
   useIntlUtils,
+  useNotification,
   useTheme,
   useTranslation,
 } from '@openmsupply-client/common';
 import { ClinicianFragment, useClinicians } from '@openmsupply-client/programs';
-import { ClinicianAutocompleteOption, Clinician } from './utils';
-import { ClinicianSlider } from './ClinicianSlider';
-import { NewClinicianModal } from './NewClinicianModal';
+import {
+  ClinicianAutocompleteOption,
+  Clinician,
+  isExistingCode,
+} from './utils';
+import { CreateClinicianSlider } from './CreateClinicianSlider';
+import { CreateClinicianModal } from './CreateClinicianModal';
+import { useCreateClinician } from './useCreateClinician';
 
 interface ClinicianSearchInputProps {
   onChange: (clinician: ClinicianAutocompleteOption | null) => void;
@@ -34,18 +41,61 @@ export const ClinicianSearchInput = ({
   mountSlidePanel = false,
 }: ClinicianSearchInputProps) => {
   const t = useTranslation();
-  const [sliderOpen, setSliderOpen] = useState(false);
-  const [modalOpen, setModalOpen] = useState(false);
-  const { data, refetch } = useClinicians.document.list({});
-  const { getLocalisedFullName } = useIntlUtils();
-  const clinicians: ClinicianFragment[] = data?.nodes ?? [];
   const theme = useTheme();
+  const { getLocalisedFullName } = useIntlUtils();
+  const { error, success } = useNotification();
+  const { data, refetch } = useClinicians.document.list({});
+  const { isSaving, draft, updateDraft, isValid, save, clear } =
+    useCreateClinician();
+
+  const clinicians: ClinicianFragment[] = data?.nodes ?? [];
+  const [modalOpen, setModalOpen] = useState(false);
+  const [sliderOpen, setSliderOpen] = useState(false);
 
   const asOption = (clinician: Clinician): ClinicianAutocompleteOption => ({
     label: getLocalisedFullName(clinician.firstName, clinician.lastName),
     value: clinician,
     id: clinician.id,
   });
+
+  const handleClinicianClose = async (clinicianId?: string) => {
+    setSliderOpen(false);
+    setModalOpen(false);
+
+    if (clinicianId) {
+      const refreshedList = await refetch();
+      const newClinician = refreshedList.data?.nodes.find(
+        c => c.id === clinicianId
+      );
+      onChange(newClinician ? asOption(newClinician) : null);
+    }
+    clear();
+  };
+
+  const handleSave = async () => {
+    try {
+      const result = await save();
+      success(t('messages.created-clinician'))();
+      handleClinicianClose(result.id);
+    } catch (e) {
+      const errorSnack = error((e as Error).message);
+      errorSnack();
+    }
+  };
+
+  const confirm = useConfirmationModal({
+    title: t('heading.are-you-sure'),
+    message: t('messages.clinician-code-already-exists'),
+    onConfirm: handleSave,
+  });
+
+  const confirmAndSave = () => {
+    if (isExistingCode(clinicians, draft.code)) {
+      confirm();
+    } else {
+      handleSave();
+    }
+  };
 
   return (
     <Box width={`${width}px`} display={'flex'} alignItems="center">
@@ -71,7 +121,11 @@ export const ClinicianSearchInput = ({
         fullWidth={fullWidth}
       />
       {allowCreate && (
-        <>
+        <Box
+          sx={{
+            overflow: 'hidden',
+          }}
+        >
           <IconButton
             icon={<PlusCircleIcon />}
             label={t('button.add-new-clinician')}
@@ -81,38 +135,28 @@ export const ClinicianSearchInput = ({
             }
           />
           {mountSlidePanel ? (
-            <ClinicianSlider
+            <CreateClinicianSlider
+              draft={draft}
+              updateDraft={updateDraft}
               width={500}
               open={sliderOpen}
-              onClose={async clinicianId => {
-                setSliderOpen(false);
-                if (clinicianId) {
-                  const refreshedList = await refetch();
-                  const newClinician = refreshedList.data?.nodes.find(
-                    c => c.id === clinicianId
-                  );
-                  onChange(newClinician ? asOption(newClinician) : null);
-                }
-              }}
-              existingClinicians={clinicians}
+              onClose={handleClinicianClose}
+              confirmAndSave={confirmAndSave}
+              isSaving={isSaving}
+              isValid={isValid}
             />
           ) : (
-            <NewClinicianModal
-              onClose={async clinicianId => {
-                setModalOpen(false);
-                if (clinicianId) {
-                  const refreshedList = await refetch();
-                  const newClinician = refreshedList.data?.nodes.find(
-                    c => c.id === clinicianId
-                  );
-                  onChange(newClinician ? asOption(newClinician) : null);
-                }
-              }}
+            <CreateClinicianModal
+              draft={draft}
+              updateDraft={updateDraft}
+              onClose={handleClinicianClose}
               open={modalOpen}
-              existingClinicians={clinicians}
+              confirmAndSave={confirmAndSave}
+              isSaving={isSaving}
+              isValid={isValid}
             />
           )}
-        </>
+        </Box>
       )}
     </Box>
   );
